// apps/wasm4pm/src/commands/init.d.ts
export declare const init: import("citty").CommandDef<{
    configFormat: {
        type: "string";
        description: string;
        alias: string;
        default: string;
    };
    force: {
        type: "boolean";
        description: string;
        alias: string;
    };
    format: {
        type: "string";
        description: string;
        default: string;
    };
    verbose: {
        type: "boolean";
        description: string;
        alias: string;
    };
    quiet: {
        type: "boolean";
        description: string;
        alias: string;
    };
}>;
//# sourceMappingURL=init.d.ts.map